Return 'trigger_uri' on assembly create and show
On assembly create, solum will generate a trigger_uri which needs to
be returned to the user. The user will then add this uri to his/her
git repo as a webhook. When a git push is done on this repo, a POST
request will be sent to the trigger_uri which will then start the
build and deploy workflow.

Change-Id: I4add3d10634f7ab4d41af948bde0bc2747d00dde