Display app_uri in app show/create responses
For now the only way to find out the URI of your
deployed assembly is to ask the assembly itself;
users interacting with Solum solely through the
'app' commands lack this important information.

Lacking the information in the plan resource, the
client searches through all the READY assemblies
built using that app ('plan') and adds the app_uri
of the most recently updated assembly to the response
of an app create or show command.

Change-Id: I5804a8002c311d714cc4cd94552635ebec5c5df8