Removing plan, component, pipeline commands from CLI display.
These can be added back, if required, later.
Currently they are confusing, as we are only planning to support
'app' and 'languagepack' as the only two nouns/concepts.
The actual removal of the code will be done when we have the
required support on the API side for these nouns.

Modified common/base to get around pep8 issue

Change-Id: I7a40a76de7a4357ced57d2701a49b4024ad3d4da
Closes-Bug: #1434665