Supporting run command
If the run_cmd is specified on the command line using a flag,
then that is used. Otherwise, the one specified in the app file
is used. User is asked to input the command if it is neither
specified in the app file nor it is specified using command line flag.

Closes-Bug: #1491624

Change-Id: I7692b7d248c77e2af2b9a19b856e2cb5f0709acb